% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment_trend.R
\name{adjustment_trend}
\alias{adjustment_trend}
\title{Adjustment to observed population trends}
\usage{
adjustment_trend(
  y_ini = c(N = 15),
  temp_cmin = 4,
  temp_cmax = 33.1,
  roi = 0.3,
  lambdai = 0.05,
  trend = 1
)
}
\arguments{
\item{y_ini}{Initial population for adjustment.}

\item{temp_cmin}{Minimum critical temperature for Desmognathus ochrophaeus (Layne & Claussen, 1987).}

\item{temp_cmax}{Maximum critical temperature for Desmognathus ochrophaeus (Markle & Kozak 2018).}

\item{roi}{Initial value for optimal intrinsic growth rate.}

\item{lambdai}{Initial value for the marginal loss of competition.}

\item{trend}{Acronym corresponding to the type of temperature trend used.}
}
\description{
This function allows to fit the observed time series of abundance and ambient
             temperature trends together with the main descriptors of the TPCs for the
             salamander Desmognathus ochrophaeus.
}
\details{
This function allows to fit the abundance trends reported for the salamander Desmognathus
        ochrophaeus in the BioTIME database (Wiley, 2016; Dornelas et al., 2018), the temperature
        data obtained from the NOAA platform (www.ncdc.noaa.gov), the descriptors of TPCs reported
        by Layne & Claussen (1987) and by Markle & Kozak (2018), using the nls2 function we can
        find the parameters ro and lambda, which are necessary to fit the abundance curve to the
        proposed model.
        Caution has to be taken to the sensitivity to the initial parameters to make adjustments,
        therefore the following temperature trends are presented in an initial proposal, the
        enumeration of which represents the option to be given in the "trend" argument of the function:
        1: Projection of increasing linear temperatura, 2: Projection of decreasing linear temperatura
        3: Periodic increasing temperature trend, 4: Periodic decreasing temperature trend,
        5: Constant temperature projection.
}
\examples{

######################################################################
  #Example 1:Fitting observed time series of abundance and ambient
             temperature, using an increasing linear temperature trend.
#######################################################################

adjustment_trend(y_ini = c(N = 15),
                temp_cmin = 4,
                temp_cmax = 33.1,
                roi = 0.3,
                lambdai = 0.05,
                trend = 1)

######################################################################
  #Example 2:Fitting the observed time series of abundance and ambient
             temperature, using an periodic increasing temperature trend.
#######################################################################

adjustment_trend(y_ini = c(N = 15),
                temp_cmin = 4,
                temp_cmax = 33.1,
                roi = 0.3,
                lambdai = 0.05,
                rend = 3)

######################################################################
  #Example 3:Fitting the observed time series of abundance and ambient
             temperature, using a constant temperature trend.
#######################################################################

adjustment_trend(y_ini = c(N = 15),
                temp_cmin = 4,
                temp_cmax = 33.1,
                roi = 0.3,
                lambdai = 0.05,
                trend = 5)


}
\references{
Dornelas M, Antão LH, Moyes F, Bates, AE, Magurran, AE, et al.(2018). BioTIME: A database of
           biodiversity time series for the Anthropocene. Global Ecol Biogeogr; 27:760 - 786.
           https://doi.org/10.1111/geb.12729.

Grothendieck, G. (2013). nls2: Non-Linear Regression with Brute Force. R package version 0.2,
           URL http://CRAN.R-project.org/package=nls2.

Layne, J. R., & Claussen, D. L. (1987). Time courses of thermal acclimation for critical
           thermal minima in the salamanders Desmognathus quadramaculatus, Desmognathus monticola,
           Desmognathus ochrophaeus, and Plethodon jordani. Comparative Biochemistry and Physiology
           Part A: Physiology, 87(4), 895–898. doi:10.1016/0300-9629(87)90011-9

Markle, T. M., & Kozak, K. H. (2018). Low acclimation capacity of narrow-ranging thermal
           specialists exposes susceptibility to global climate change. Ecology and Evolution, 8(9),
           4644–4656.doi:10.1002/ece3.4006.

NOAA National Centers for Environmental Information, State of the Climate: Global Climate
           Report for Annual 2020, published online January 2021, retrieved on September 4, 2021 from
           https://www.ncdc.noaa.gov/sotc/global/202013.

Wiley, R. H. (2016). “Population estimates of Appalachian salamanders”. Coweeta LTER. Available
           at: http://coweeta.uga.edu/eml/1044.xml.
}
