% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_structure.R
\name{age_structure}
\alias{age_structure}
\title{Age structure under IPCC RCP2.6 or RCP8.5 scenarios}
\usage{
age_structure(
  y_ini = c(N1 = 800, N1 = 800, N1 = 800, N2 = 600, N2 = 600, N2 = 600, N3 = 400, N3 =
    400, N3 = 400),
  temp_ini = rep(25, 3),
  temp_cmin = rep(18, 3),
  temp_cmax = c(25, 28, 35),
  ro = rep(0.7, 3),
  lambda1 = rep(4e-04, 3),
  lambda2 = rep(4e-04, 3),
  lambda3 = rep(4e-04, 3),
  alpha1 = 0.1,
  alpha2 = 0.7,
  d2 = 0.4,
  d3 = 0.4,
  RCP = 2.6,
  time_start = 2005,
  time_end = 2100,
  leap = 1/50
)
}
\arguments{
\item{y_ini}{Initial population values (must be written with its name: N).}

\item{temp_ini}{Initial temperature.}

\item{temp_cmin}{Minimum critical temperature (prey).}

\item{temp_cmax}{Maximum critical temperature (prey).}

\item{ro}{Population growth rate at optimum temperature (prey).}

\item{lambda1}{Marginal loss by non-thermodependent intraspecific competition (age1).}

\item{lambda2}{Marginal loss by non-thermodependent intraspecific competition (age2).}

\item{lambda3}{Marginal loss by non-thermodependent intraspecific competition (age3).}

\item{alpha1}{Stage 1 to 2 transition coefficient.}

\item{alpha2}{Stage 2 to 3 transition coefficient.}

\item{d2}{Mortality rate (age2).}

\item{d3}{Mortality rate (age3).}

\item{RCP}{Representative concentration trajectories (RCP2.6 and RCP8.5 scenarios).}

\item{time_start}{Start of time sequence.}

\item{time_end}{End of time sequence.}

\item{leap}{Time sequence step.}
}
\value{
(1) A data.frame with columns having the simulated trends.

(2) A four-panel figure where (a), (b), and (c) show the population abundance curves for each simulation.
       The brown curve corresponds to the abundance of prey and the green curve to predators. Panel (d)
       shows the temperature trend curves used for each simulation, green, blue, and black, respectively.
}
\description{
This function allows simulating the effect of the IPCC RCP2.6 or RCP8.5 scenarios
            (2014) on the abundances of two species interacting through predation. The prey is an
            ectotherm population, and the predator is not affected by temperature.
}
\details{
Three scenarios can be evaluated for a predation interaction where the prey is an ectotherm
        population. The temperature trends correspond to IPCC projections under the RCP2.6 or RCP8.5
        scenarios. In each input vector, the parameters for the three simulations must be specified.
}
\examples{

#######################################################################
  #Example 1: Different thermal tolerance ranges (scenario RCP2.6).
#######################################################################

temp_cmin <- 18

# Temperature that occurs before the minimum simulation time.
temp_i <- 22

time_end <- 2100

# Temperature that occurs in the maximum time ofthe simulation.
temp_max <- get_RCP2.6(time_end)+temp_i

# Simulation thermal range.
RS <- temp_max-temp_cmin

temp_cmax1 <- 4/3*RS+temp_cmin
temp_cmax2 <- 2/3*RS+temp_cmin
temp_cmax3 <- 1/3*RS+temp_cmin
temp_ini <- (temp_cmin+temp_cmax3)/2

age_structure(y_ini = c(N1 = 800, N1 = 800, N1 = 800,
                       N2 = 600, N2 = 600, N2 = 600,
                       N3 = 400, N3 = 400, N3 = 400),
             temp_ini = rep(temp_ini,3),
             temp_cmin = rep(temp_cmin,3),
             temp_cmax = c(temp_cmax1,temp_cmax2,temp_cmax3),
             ro = rep(0.7,3),
             lambda1 = rep(0.0004,3),
             lambda2 = rep(0.0006,3),
             lambda3 = rep(0.0005,3),
             alpha1 = 0.3,
             alpha2 = 0.4,
             d2 = 0.2,
             d3 = 0.3,
             RCP = 2.6,
             time_start = 2005,
             time_end = time_end,
             leap = 1/50)

#######################################################################
  #Example 2: Different thermal tolerance ranges (scenario RCP8.5).
#######################################################################

temp_cmin <- 18

# Temperature that occurs before the minimum simulation time.
temp_i <- 22

time_end <- 2100

# Temperature that occurs in the maximum time ofthe simulation.
temp_max <- get_RCP8.5(time_end,a=exp(coef(m)[1]), b=coef(m)[2])+temp_i

# Simulation thermal range.
RS <- temp_max-temp_cmin

temp_cmax1 <- 4/3*RS+temp_cmin
temp_cmax2 <- 2/3*RS+temp_cmin
temp_cmax3 <- 1/3*RS+temp_cmin
temp_ini <- (temp_cmin+temp_cmax3)/2

age_structure(y_ini = c(N1 = 800, N1 = 800, N1 = 800,
                       N2 = 600, N2 = 600, N2 = 600,
                       N3 = 400, N3 = 400, N3 = 400),
             temp_ini = rep(temp_ini,3),
             temp_cmin = rep(temp_cmin,3),
             temp_cmax = c(temp_cmax1,temp_cmax2,temp_cmax3),
             ro = rep(0.7,3),
             lambda1 = rep(0.0004,3),
             lambda2 = rep(0.0006,3),
             lambda3 = rep(0.0005,3),
             alpha1 = 0.3,
             alpha2 = 0.4,
             d2 = 0.3,
             d3 = 0.4,
             RCP = 8.5,
             time_start = 2005,
             time_end = time_end,
             leap = 1/50)

#######################################################################
  #Example 3: Different conversion efficiencies (scenario RCP2.6).
#######################################################################

e1 <- 0.2
e2 <- 2*e1
e3 <- 2*e2

age_structure(y_ini = c(N1 = 800, N1 = 800, N1 = 800,
                       N2 = 600, N2 = 600, N2 = 600,
                       N3 = 400, N3 = 400, N3 = 400),
             temp_ini = rep(22,3),
             temp_cmin = rep(20,3),
             temp_cmax = rep(35,3),
             ro = rep(0.9,3),
             lambda1 = rep(0.0004,3),
             lambda2 = rep(0.0006,3),
             lambda3 = rep(0.0005,3),
             alpha1 = 0.3,
             alpha2 = 0.4,
             d2 = 0.3,
             d3 = 0.4,
             RCP = 2.6,
             time_start = 2005,
             time_end = 2100,
             leap = 1/50)

#######################################################################
  #Example 4: Different conversion efficiencies (scenario RCP8.5).
#######################################################################

e1 <- 0.2
e2 <- 2*e1
e3 <- 2*e2

age_structure(y_ini = c(N1 = 800, N1 = 800, N1 = 800,
                       N2 = 600, N2 = 600, N2 = 600,
                       N3 = 400, N3 = 400, N3 = 400),
             temp_ini = rep(22,3),
             temp_cmin = rep(20,3),
             temp_cmax = rep(35,3),
             ro = rep(0.9,3),
             lambda1 = rep(0.0004,3),
             lambda2 = rep(0.0006,3),
             lambda3 = rep(0.0005,3),
             alpha1 = 0.3,
             alpha2 = 0.4,
             d2 = 0.3,
             d3 = 0.4,
             RCP = 8.5,
             time_start = 2005,
             time_end = 2100,
             leap = 1/50)
}
\references{
IPCC. (2014): Cambio climático 2014: Informe de síntesis. Contribución de los Grupos de trabajo
           I, II y III al Quinto Informe de Evaluación del Grupo Intergubernamental de Expertos sobre el
           Cambio Climático [Equipo principal de redacción, R.K. Pachauri y L.A. Meyer (eds.)]. IPCC, Ginebra,
           Suiza, 157 págs.
}
