% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epcc-package.R
\docType{package}
\name{epcc-package}
\alias{epcc}
\alias{epcc-package}
\title{epcc: Simulating Populations of Ectotherms under Global Warming}
\description{
Provides several functions that allow model and simulate 
             the effects of thermal sensitivity and the exposition to 
             different trends in environmental temperature on the 
             abundance dynamics of ectotherms populations. It allows 
             an easy implementation of the possible consequences of 
             warming at global and local scales, constituting a useful 
             tool for understanding the extinction risk of populations.
             (Víctor Saldaña-Núñez, Fernando Córdova-Lepe, & Felipe N. 
             Moreno-Gómez, 2021) <doi:10.5281/zenodo.5034087>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Victor-Saldana/epcc}
  \item Report bugs at \url{https://github.com/Victor-Saldana/epcc/issues}
}

}
\author{
\strong{Maintainer}: Víctor Saldaña-Núñez \email{saldananunezvictor@gmail.com}

Authors:
\itemize{
  \item Fernando Córdova-Lepe
  \item Felipe N. Moreno-Gómez
}

}
\keyword{internal}
