% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epcc-package.R
\docType{package}
\name{epcc-package}
\alias{epcc}
\alias{epcc-package}
\title{epcc: epcc: an R package to assess and simulate the vulnerability of ectothermic populations when facing global warming}
\description{
This package provides several functions 
             that allow model and simulate the effects of thermal 
             sensitivity and the exposition to different trends in  
             environmental temperature on the abundance dynamics of 
             ectothermic populations. It allows an easy 
             implementation of the possible consequences of warming 
             at global and local scales, constituting a useful tool 
             for understanding the extinction risk of populations.
}
\keyword{internal}
