% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RCP8.5.R
\name{get_RCP8.5}
\alias{get_RCP8.5}
\title{Projected values under IPCC RCP8.5 scenario}
\usage{
get_RCP8.5(date, a, b)
}
\arguments{
\item{date}{A specific year or a vector of years.}

\item{a, b}{Parameters obtained from the adjustment of the temperature trend.}
}
\description{
This function allows obtaining the projected increment in environmental
            temperature according to the IPCC RCP8.5 scenario.
}
\details{
The temperature increment projection of the change in global mean surface
        temperature according to the IPCC RCP8.5 scenario. It is possible to get
        the value for one or various years.
}
\examples{

########################################################################
 #Example 1: Projection of the temperature increase for a given year.
########################################################################

date <- 2050
temp <- get_RCP8.5(date,a=exp(coef(m)[1]), b=coef(m)[2])
temp

########################################################################
  #Example 2: Projection of the temperature increase for a vector of years.
########################################################################

date <- seq(2005,2100,1/12)
temp <- get_RCP8.5(date,a=exp(coef(m)[1]), b=coef(m)[2])
plot(date,temp,type="l")

}
\references{
IPCC. (2014): Cambio climático 2014: Informe de síntesis. Contribución de los Grupos de trabajo
           I, II y III al Quinto Informe de Evaluación del Grupo Intergubernamental de Expertos sobre el
           Cambio Climático [Equipo principal de redacción, R.K. Pachauri y L.A. Meyer (eds.)]. IPCC, Ginebra,
           Suiza, 157 págs.
}
